+++ src/main/java/com/marmitech/Marmitech/Repository/UsuarioRepository.java
@@ -3,5 +3,5 @@ package com.marmitech.Marmitech.Repository;
 import com.marmitech.Marmitech.Model.UsuarioEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Integer> {
+public interface UsuarioRepository extends JpaRepository<UsuarioEntity, Long> {
 }
